Add from_beginning options to consumer group
The from_beginning option allow you to define the consuming
strategy for a consumer group. If it's set to true, the consumer
will start to consume from the offset next to the last commited one.
If it's set to false, the consumer will start to consume next to
the last offset.