Version control for each protocol' API
For each API of the protocol, we added the max supported version. If, in
the configuration, the setted version is higher than the max supported
version (MAX_VERSION) by the API, we use the MAX_VERSION.

For example, the current Offset API max version is 1. So if in the
configuration we set 2 as the version to use, we will use the Offset API
version 1. If in the configuration we set 0 or 1 as the version to use,
we will, respectively, use the Offset API version 0 or 1.