Do not get stuck waiting for a leader to become available
If any partition had an unavailable leader, kafe_brokers would
get into a loop of update_state_with_metadata/2 and become
unresponsive until the cluster had found a leader, even though
it might concern a topic we were not interested in.

The state update now always proceeds to the end of the topic list,
leaving a leader_not_available atom instead of the leader ID when
a leader is not available.

Also, the error tuple when a partition leader is not available when
fetching or producing is now more explicit with a "leader_not_available"
reason.

Keep partitions which don't have a leader in metadata

This avoids an edge case where a partition would seem to not exist
if it temporarily had no leader available.